requirements.txt
tensorflow
keras
pandas
numpy
streamlit
pillow
plotly